+++ package.json
@@ -14,11 +14,13 @@
   "dependencies": {
     "bcrypt": "^3.0.7",
     "body-parser": "^1.19.0",
+    "compression": "^1.7.4",
     "dotenv": "^8.2.0",
     "ejs": "^3.0.1",
     "express": "^4.17.1",
     "express-flash": "0.0.2",
     "express-session": "^1.17.0",
+    "helmet": "^3.21.2",
     "method-override": "^3.0.0",
     "mongoose": "^5.8.3",
     "nodemailer": "^6.4.2",
