+++ src/auth-helper.js
@@ -1,0 +1,19 @@
+const checkAuthenticated = (req, res, next) => {
+    if (req.isAuthenticated()) {
+        return next()
+    }
+
+    res.redirect('/login')
+};
+  
+const checkNotAuthenticated = (req, res, next) => {
+    if (req.isAuthenticated()) {
+        return res.redirect('/guests')
+    }
+    next()
+};
+
+module.exports = {
+    checkAuthenticated,
+    checkNotAuthenticated
+};
