+++ src/routes/home.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+// GET - home page (will redirect to login).
+router.get("/", (req, res) => {
+    res.redirect("/login");
+});
+
+module.exports = router;
